apps/users/serializers.py
from django.contrib.auth import get_user_model
from rest_framework import serializers
from .models import TeacherProfile

User = get_user_model()


class RegisterSerializer(serializers.ModelSerializer):
    password = serializers.CharField(write_only=True, min_length=8)

    class Meta:
        model = User
        fields = ("id", "username", "email", "password", "role")

    def create(self, validated_data):
        password = validated_data.pop("password")
        user = User(**validated_data)
        user.set_password(password)
        user.save()
        return user


class MeSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ("id", "username", "email", "role")

class TeacherProfileSerializer(serializers.ModelSerializer):
    username = serializers.SerializerMethodField(source="get_username")
    email = serializers.SerializerMethodField(source="get_email")
    class Meta:
        model = TeacherProfile
        fields = [
            "id",
            "username",
            "email",
            "avatar",
            "bio",
            "specialization",
        ]