requirements/base.txt
Django>=5.0
djangorestframework
python-decouple
djangorestframework-simplejwt